Make consumption function return Void
Our Nakadi instance experienced a data loss and thus started reporting
negative unconsumed event counts.
When a consumption is started in this state Nakadi will return a 200
with an empty body.

Currently this simply terminated the conduit @linesUnboundedAsciiC@.
With this change this will throw a @ConsumptionStoppedException@.

I change the return type to Void to signal that the consumption function
should never terminate (in our codebase people simply wrapped it in an
infinite loop).
The Nakadi docs (https://nakadi.io/manual.html#consuming-events-from-a-subscription)
only talk about an "info" field returned in batches,
which can "potentialy contain the reason Nakadi closed the connection".
No more information about Nakadi-initiated closing of connections is
provided.